move magic into chess module
