+++ src/chess/mod.rs
@@ -2,6 +2,7 @@ use std::sync::atomic::AtomicBool;
 
 pub mod board;
 pub mod chessmove;
+mod magic;
 pub mod piece;
 pub mod piecelayout;
 pub mod squareset;
